refactor(showcase): change file names to showcase
